+++ src/holmes-rs/src/hydro/mod.rs
@@ -2,6 +2,7 @@ use pyo3::prelude::*;
 pub mod bucket;
 pub mod cequeau;
 pub mod crec;
+pub mod gardenia;
 pub mod gr4j;
 pub mod utils;
 use crate::utils::register_submodule;
@@ -15,6 +16,12 @@ pub fn make_module(py: Python<'_>) -> PyResult<Bound<'_, PyModule>> {
     register_submodule(py, &m, &bucket::make_module(py)?, "holmes_rs.hydro")?;
     register_submodule(py, &m, &cequeau::make_module(py)?, "holmes_rs.hydro")?;
     register_submodule(py, &m, &crec::make_module(py)?, "holmes_rs.hydro")?;
+    register_submodule(
+        py,
+        &m,
+        &gardenia::make_module(py)?,
+        "holmes_rs.hydro",
+    )?;
     Ok(m)
 }
 
@@ -26,6 +33,7 @@ pub fn get_model(
         "bucket" => Ok((bucket::init, bucket::simulate)),
         "cequeau" => Ok((cequeau::init, cequeau::simulate)),
         "crec" => Ok((crec::init, crec::simulate)),
+        "gardenia" => Ok((gardenia::init, gardenia::simulate)),
         _ => Err(HydroError::WrongModel(model.to_string())),
     }
 }