+++ src/holmes-rs/tests/unit/hydro/mod.rs
@@ -1,4 +1,5 @@
 mod bucket_tests;
 mod cequeau_tests;
 mod crec_tests;
+mod gardenia_tests;
 mod gr4j_tests;